+++ backend/src/main/java/org/example/repository/CustomerAddressRepository.java
@@ -3,4 +3,4 @@ package org.example.repository;
 import org.example.model.CustomerAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerAddressRepository extends JpaRepository<CustomerAddress, Long> { }
+public interface CustomerAddressRepository extends JpaRepository<CustomerAddress, Integer> { }
